+++ libs/identity/feature-access-status/src/lib/identity-feature-access-status/identity-feature-access-status.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'lib-identity-feature-access-status',
+  imports: [],
+  templateUrl: './identity-feature-access-status.html',
+  styleUrl: './identity-feature-access-status.scss',
+})
+export class IdentityFeatureAccessStatus {}