+++ libs/identity/feature-greeting/src/index.ts
@@ -1,0 +1,1 @@
+export { GreetingPageComponent } from './lib/greeting-page/greeting-page';