+++ libs/identity/feature-greeting/src/lib/identity-feature-greeting/identity-feature-greeting.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'lib-identity-feature-greeting',
+  imports: [],
+  templateUrl: './identity-feature-greeting.html',
+  styleUrl: './identity-feature-greeting.scss',
+})
+export class IdentityFeatureGreeting {}